# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Rental Car POS & Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List all vehicles
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [available, rented, maintenance, null]
        - name: ownership
          in: query
          schema:
            type: ["string", "null"]
            enum: [internal, external, null]
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Add a new vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleBody"
      responses:
        "201":
          description: Created vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get a vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Not found
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleBody"
      responses:
        "200":
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Delete a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── MAINTENANCE ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        plateNumber:
          type: string
        year:
          type: integer
        color:
          type: string
        status:
          type: string
          enum: [available, rented, maintenance]
        ownership:
          type: string
          enum: [internal, external]
        partnerName:
          type: ["string", "null"]
        dailyRate:
          type: number
        profitSharePercent:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - plateNumber
        - year
        - color
        - status
        - ownership
        - partnerName
        - dailyRate
        - profitSharePercent
        - notes
        - createdAt
    CreateVehicleBody:
      type: object
      properties:
        name:
          type: string
        plateNumber:
          type: string
        year:
          type: integer
        color:
          type: string
        status:
          type: string
          enum: [available, rented, maintenance]
        ownership:
          type: string
          enum: [internal, external]
        partnerName:
          type: ["string", "null"]
        dailyRate:
          type: number
        profitSharePercent:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - plateNumber
        - year
        - color
        - ownership
        - dailyRate
    UpdateVehicleBody:
      type: object
      properties:
        name:
          type: string
        plateNumber:
          type: string
        year:
          type: integer
        color:
          type: string
        status:
          type: string
          enum: [available, rented, maintenance]
        ownership:
          type: string
          enum: [internal, external]
        partnerName:
          type: ["string", "null"]
        dailyRate:
          type: number
        profitSharePercent:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
...